backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.8.2
pydantic-settings==2.3.4
langchain==0.2.11
langchain-core==0.2.22
langchain-community==0.2.9
langchain-google-genai==2.1.9
google-ai-generativelanguage>=0.6.18,<0.7.0
langchain-mongodb==0.1.7
pymupdf==1.24.9
pymongo==4.8.0
supabase==2.5.0
python-jose[cryptography]==3.3.0
pytest==8.3.2
httpx==0.27.0